adds more tests
- refactor tests
- add enc-dec tests
- skips generate tests for non-lora adapters
